feat: :fire: comecando estilizar dash

+++ system/database/migrations/tenant/2021_09_28_212629_add_completar_cadastro_field_to_empresa.php
@@ -14,7 +14,7 @@ class AddCompletarCadastroFieldToEmpresa extends Migration
     public function up()
     {
         Schema::table('empresa', function (Blueprint $table) {
-            $table->tinyInteger('completar_cadastro')->after('tipo_telefone');
+            $table->integer('completar_cadastro')->after('tipo_telefone');
         });
     }
